Move metrics registry and initialization to pkg/metricsconfig
Before, metrics registry was defined in pkg/metrics and initialization
functions in pkg/metrics/metricsconfig. It make sense to keep them together and
outside of pkg/metrics, which contains a library for defining metrics. Now all
code related to running the metrics server sits in pkg/metricsconfig.
