+++ pkg/metricsconfig/root.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Tetragon
 
-package metrics
+package metricsconfig
 
 import (
 	"net/http"
